Did some support for D3D wrapper by mh, original source you can find at http://quakeone.com/mh/ Unfortunately ezquake require a bit more OpenGL functionality than GLQuake, so there some issues. I have no time to fix it ATM so added empty stubs for:
glCopyTexSubImage2D (used for bloom only), glRectf (used for menus and such, not a problem to fix it), glLineStipple (sss experimental code, not a problem to fix and not actually critical), glPushAttrib/glPopAttrib (may be fixed too), glGetError (not a critical at all), few more stubs but they are not really used by ezquake but rather just defined.

It all was done just to test 3D 120hz monitors in stereo mode since they artificially limited to the D3D only, world wide conspiracy :(

Ah, and you require DirectX 8 headers/libs, I used ones which supplied with http://quakeone.com/mh/dx8fuh.zip, that again fault of MS that dx8 is not comming with latest DirectX SDK, but perhaps it would be nice to change sources to use dx9 instead of dx8.