Added scr_scoreboard_spectator_name
Allows changing what spectators are called in the scoreboard. When
teamplay is not on, this variable is cut to 4 characters only. The old
multi-colored spectator code should probably be included in the changelog
in case anyone wants to use it. Default is brown spectator. Huge thanks to
Johnny_cz for helping cut the variable length