Fix lots of leaks with call stacks under sv_mod_msg_file_OnChange
When using the built-in list of qw messages there is no need for allocation.
The alloc was immediately leaked when we point to the built-in message.

Also, there already exists a mechanism (qwm_static) to avoid accidentally
attempting to free the built-in messages if the file changes.

Leak: 0x7fd3f254b560  size=32  zone: DefaultMallocZone_0x1107c9000
	0x00000000 0x00000000 0x00000000 0x00000000 	................
	0x00000000 0x00000000 0x00000000 0x00000000 	................
	Call stack: [thread 0x7fff79ffb310]:
        | 0x1
        | start
        | main sys_posix.c:310
        | Host_Init host.c:625
        | Cvar_Set cvar.c:310
        | sv_mod_msg_file_OnChange sv_mod_frags.c:67
        | Q_malloc q_shared.c:884
        | malloc
        | malloc_zone_malloc