DOWNLOADING: Redo downloading to make it work on FTE servers
FTE servers send modellist before we ask for it so we need to be able to
cope with that, otherwise it will trash our sound downloads ...

First get the lists now and when we have soundlist and modellist, then
start with downloads.

Remove the renaming boolean passing and instead rely on
cls.downloadpercent (and actually set that one properly).