cl_fakeshaft 3: experimental feature to assist high-ping players using LG.
cl_fakeshaft_clientside_beams:
  if enabled & fakeshaft 3, client will create current player's beam and
     sound effects.  This should remove delay when initially firing LG,
     and delay when stopping.
cl_fakeshaft_clientside_damage:
  if enabled & fakeshaft 3, client will predict if the player's beam has
     hit another player, and spawn blood to indicate if so.  blood coming
     from the server is ignored if sufficiently close to the last hit-point,
     otherwise allowed through.

Client-side damage can only be predicted against other players, so the
standard delay will be visible if (for instance) firing against the LG door
on DM6, or the button on DM2.

Not suitable for non-standard DM QW - it will always assume that weapon 8
is a standard LG.  Also on low ping it is probably better to stay on
cl_fakeshaft 2, as the response from the server will be fast enough to
make no difference, and the player will receive full confidence of damage
inflicted.

We do not currently predict cell usage (so ammo reduction will be delayed)
or damage sounds from the enemy players (so the player may see blood from
hitting the player but their model changing & sound effects will be delayed).

This is very experimental as we are essentially giving feedback to the
player that the server has not confirmed.