Added support for multi-monitor systems
Can now choose which display to use, through vid_displayNumber &
vid_win_displayNumber (for fullscreen & windowed modes accordingly).  If
a display is not available then the system should default back to the main
display for the system.  Can also select using -display on command line.

Co-ordinates are stored relative to the display they are on, so if you have
a window position stored on display 1, changing the resolution of display 0
should have no impact.