Revert "cl_fakeshaft 3: experimental feature to assist high-ping players using LG."
This reverts commit a6870fdec9a82e42402ce112a660b68511dc456a.

After playing online with this, there are too many issues with tracking
the current user's weapon in certain situations, and also with correctly
filtering the blood from the server (the network protocol doesn't tell us
what beam caused the damage or by who, so we go by distance, which doesn't
work when the enemy target is moving quickly, including when we are
pushing them back) that the actual results aren't good enough yet.

A lot more thought needs to go into it before it is ready for public
consumption.  Reverting for the moment, will look at again in the future.