VID_SDL2: Add temporary gamma workaround for Linux systems
There's an issue with SDL2 and Xorg regarding gamma making it not work
properly on some setups. So for the moment lets use the native X11 API's
to get/set gamma instead of SDL :/

Quickly tested, might need some fixing up and some refactoring could be
done to make it prettier.