VIDEO: Fix conwidth/conheight on cmdline and more
Add /modelist command to list supported resolutions in console

Add a cvar setter that acts just as Cvar_Set but won't trigger callback
(using this to set values to cvars that are set on cmdline (we need
information that yet isn't initialized so triggering callback would set
default values...)

vid_displayfrequency 0 (and vid_usedesktopres 0) would result in no
video mode match. Fix so that it returns the mode with highest refresh
rate if width/height matches.