HUD_COMMON/GL_DRAW: Ugly tmp fix for a hud_groupX vid_restart problem
The picture data is loaded into a Cache that gets wiped on vid_restart.
The hud_pic_groupX pointers point to invalid data and doesn't get reset
so just force reset these when the cache is cleared to force a reload.

All of this needs a major overhaul one day...