MULTIVIEW: Don't reset variables on initial startup
A number of variables need to be set to sensible values during multiview
playback.  If not in multiview, the old values are saved, and then restored
before initiating a new connection.  If the command line caused an
immediate connection, the values would be set to their default values,
ignoring options that the user has set in their config file.