LEAK: Fix leak during tab completion (escape_regex)
All calls to escape_regex leak. It allocates a string (double the size of
the input), returns it, and it is never freed. Since it is only called
in a few places use a local and ensure we free.

To reproduce:

    1. Open console
    2. Type "src" <tab>
      => Leaks

Leak: 0x7fe742c44760  size=16  zone: DefaultMallocZone_0x108d1a000  length: 3  "scr"
	Call stack: [thread 0x7fff7a6a9000]:
        | start
        | main sys_posix.c:340
        | Host_Frame host.c:465
        | CL_Frame cl_main.c:2446
        | Sys_SendKeyEvents vid_sdl2.c:272
        | IN_Frame vid_sdl2.c:610
        | Key_Event keys.c:2250
        | Key_EventEx keys.c:2198
        | CompleteCommandNew keys.c:2323
        | Q_malloc q_shared.c:884
        | malloc
        | malloc_zone_malloc