LEAK: Fix leak during tab completion (str_repeat)
All calls to str_repeat leak. It can allocate a string and return it,
and it is never freed. Since it is only called in a few places use a
local and ensure we free.

Cleanup the code by making a helper to print the mark and padding.
This simplifies PaddedPrintValue and makes it easy to see the
lifetime of this allocation.

To reproduce:

    1. Open console
    2. Type "src" <tab>
      => Leaks

Leak: 0x7fe742d545d0  size=16  zone: DefaultMallocZone_0x108d1a000  length: 2  "  "
	Call stack: [thread 0x7fff7a6a9000]:
        | start
        | main sys_posix.c:340
        | Host_Frame host.c:465
        | CL_Frame cl_main.c:2446
        | Sys_SendKeyEvents vid_sdl2.c:272
        | IN_Frame vid_sdl2.c:610
        | Key_Event keys.c:2250
        | Key_EventEx keys.c:2198
        | CompleteCommandNew keys.c:570
        | PaddedPrintValue keys.c:365
        | str_repeat utils.c:65
        | Q_calloc q_shared.c:898
        | calloc
        | malloc_zone_calloc