VX_TRACKER: Update r_tracker_inconsole with two new options
r_tracker_inconsole 1 disabled regular tracker but printed in console,
this buggy behaviour is kept but two new options introduced.

Allow for tracker messages to be printed in console but also print to
the regular tracker area.

Last option prints to console and tracker area but it won't be shown in
the notification area

Request from Milton