INTERNAL SERVER: Restore client-side pmove struct over SV_Frame
This resolves issue with 'stuttering' during movement when using internal
server and cl_delay_packet.

client-side pmove struct contains physical/solid entities as per the last
received packet, and is set when the packet is read from the server.  when
using cl_delay_packet that code is not run immediately after SV_Frame,
leaving pmove to be state set during server processing of user commands.

This is especially bad when running bots, as the pmove struct will contain
the current player for clipping during the bot's move, leading to pmove
logic thinking the current player is blocked inside another solid entity,
cl.simvel will be [0 0 0] and the stuttering will be very severe.