BUILD: remove build dependency on xxd utility
The xxd utility is included with Vim.  It is used in the build
process to convert JSON files into a representation suitable for
inclusion in a C program.

The same task can be accomplished with just regular POSIX
utilities -- mainly awk and od -- which are already present
on most operating systems.  This avoids having to install an
additional complete editor just for this small task in the build
process.