FEAT: Add easymove from jquake
Easymove is a mode where you can strafejump simply by holding shift.

It is usable by people who can't bunny jump properly because of
e.g. a disability or similar, and helps them to be competitive
in games with friends.

It is disabled by default in ruleset smackdown.

The feature was written by Tonik.