Fix build on FreeBSD/powerpc
While the code was fixed for powerpc64 (and powerpc64le), it still fails to build on powerpc (32-bit). Since it's possible to use the same way on 32-bits, do it to fix build.
It may fail on the early powerpc chips, but I doubt they are able to run a modern OS with OpenGL anyway.