FS: Remove last traces of iscompressed in vfs_zip.
Most of the uses had already been removed, but remained in the case of a
seek. As cmodel now uses seeks during checksum, this uncovered the last
use case.

As iscompressed wasn't set, the vfsz->defer remained NULL, and thus
reading of pk3 content after a seek failed.

The origin of this file is FTE, ideally a newer version is imported but
the rest of the FTE VFS guts has evolved since last import.

fixes #885