STUFFTEXT: Add set_ktx
For user convenience, there's a new command called `set_ktx` that sets
`cl_stufftext_allowed_commands` to the values that allow KTX to work
properly. However, aliases and downloads are disabled for security
reasons.

All commands defined in the KTX source code are also loaded as
`tempaliases` to make it easier for users to maintain security while
still being able to use KTX almost in the same way as before.