STYLE: Fix some c89 violations.
The wiki specifies that C89 is the target standard. In reality
the codebase adheres to GNU89. When actually enforcing this the
behavior of MinGW cross-compilation changes, which allows
JSON_INTEGER_FORMAT to work as intended.

Later standards automatically enable __USE_MINGW_ANSI_STDIO=1
in MinGW which is the cause of the jansson breakage.

It seems to be fairly common to combine __STDC_FORMAT_MACROS=1
and __USE_MINGW_ANSI_STDIO=0 in projects that disable it.