Add port ping probe
By setting cl_portpingprobe_enable to 1, your client will perform
cl_portpingprobe_probes number of probes on the target server before
establishing the connection.

This means that a connectionless ping will be sent to the target server
from different source ports to measure which source port that gives the
lowest latency, selecting that port as the source for the connection.

When using a proxy, the proxy server will be probed instead. This
feature works for both connect and connectbr.

The new variables are:

- cl_portpingprobe_enable (default: 0)
- cl_portpingprobe_probes (default: 500)
- cl_portpingprobe_delay (default: 0)