Execute TP triggers in the appropriate context (#1075)
A malicious proxy or server can inject a f_trigger alias that gets
executed in the cbuf main context, effectively circumventing the remote
capabilities check and enabling remote code execution.

This fix checks whether the alias was submitted by a server, and if so,
executes it in the server context instead.